Added FLV support in FLU file requester. Added UI color support in preferences (albeit it refreshes only when saved).
Former-commit-id: 3003d63897b54ff278e2788ec50be8f03cefa588

+++ mrViewer/src/standalone/mrvCommandLine.h
@@ -28,7 +28,7 @@ typedef std::vector< std::string > stringArray;
        float gain;
        std::string host;
        stringArray audios;
-       unsigned port;
+       unsigned short port;
        float fps;
 
   Options() : edl(false), gamma(1.0f), gain( 1.0f ), port( 0 ), fps( 0 )
